+++ src/main/java/com/banking/springbootbanking/service/TransactionService.java
@@ -4,6 +4,7 @@ import com.banking.springbootbanking.model.Transaction;
 import com.banking.springbootbanking.model.dto.TransactionDTO;
 import com.banking.springbootbanking.utils.enums.CurrencyType;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface TransactionService {
@@ -11,17 +12,17 @@ public interface TransactionService {
 
     TransactionDTO getTransactionById(Long id);
 
-    TransactionDTO accountWithdraw(String number, Float amount, CurrencyType currency);
+    TransactionDTO accountWithdraw(String number, BigDecimal amount, CurrencyType currency);
 
-    TransactionDTO cardWithdraw(String number, Float amount, CurrencyType currency);
+    TransactionDTO cardWithdraw(String number, BigDecimal amount, CurrencyType currency);
 
-    TransactionDTO accountDeposit(String number, Float amount, CurrencyType currency);
+    TransactionDTO accountDeposit(String number, BigDecimal amount, CurrencyType currency);
 
-    TransactionDTO cardDeposit(String number, Float amount, CurrencyType currency);
+    TransactionDTO cardDeposit(String number, BigDecimal amount, CurrencyType currency);
 
-    TransactionDTO accountTransfer(String senderNumber, String recipientNumber, Float amount, String description, CurrencyType currency);
+    TransactionDTO accountTransfer(String senderNumber, String recipientNumber, BigDecimal amount, String description, CurrencyType currency);
 
-    TransactionDTO cardTransfer(String senderNumber, String recipientNumber, Float amount, String description, CurrencyType currency);
+    TransactionDTO cardTransfer(String senderNumber, String recipientNumber, BigDecimal amount, String description, CurrencyType currency);
 
     List<TransactionDTO> getAllTransactions();
 }
